requirements.txt
# Core Dependencies
python-telegram-bot==20.7
python-dotenv==1.0.0

# Database
SQLAlchemy==2.0.23
alembic==1.13.0

# HTTP Requests (for Yandex GPT API)
requests==2.31.0
aiohttp==3.9.1

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0

# Code Quality
black==23.12.1
pylint==3.0.3
mypy==1.7.1

# Utilities
pydantic==2.5.2
pydantic-settings==2.1.0
